+++ tools.cpp
@@ -104,8 +104,8 @@ std::set<GeometricObject*> TangentLineCreator::apply(std::vector<GeometricObject
     }
     Circle* circle = dynamic_cast<Circle*>(objs[1]);
     QPointF p1 = objs[0]->position(), p2 = circle->position();
-    double radius = circle->getRadius();
-    double dist = std::sqrt(std::pow(p1.x() - p2.x(), 2) + std::pow(p1.y() - p2.y(), 2));
+    long double radius = circle->getRadius();
+    long double dist = std::sqrt(std::pow(p1.x() - p2.x(), 2) + std::pow(p1.y() - p2.y(), 2));
     if (radius > dist + 1e-4) {
         return std::set<GeometricObject*>();
     } else if (abs(radius - dist) <= 1e-4) {
